#10070: Move log2 bw to TTNN
