+++ docs/source/ttnn/ttnn/ttnn/log2_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.log2_bw:
+
+ttnn.log2_bw
+###############
+
+.. autofunction:: ttnn.log2_bw
