+++ tests/ttnn/unit_tests/operations/backward/test_backward_log2.py
@@ -4,8 +4,8 @@
 
 import torch
 import pytest
-import tt_lib
-from tests.tt_eager.python_api_testing.unit_testing.backward_ops.utility_funcs import (
+import ttnn
+from tests.ttnn.unit_tests.operations.backward.utility_funcs import (
     compare_pcc,
     data_gen_with_range,
 )
@@ -23,7 +23,7 @@ def test_bw_log2(input_shapes, device):
     in_data, input_tensor = data_gen_with_range(input_shapes, -100, 100, device, True)
     grad_data, grad_tensor = data_gen_with_range(input_shapes, -10, 10, device)
 
-    tt_output_tensor_on_device = tt_lib.tensor.log2_bw(grad_tensor, input_tensor)
+    tt_output_tensor_on_device = ttnn.log2_bw(grad_tensor, input_tensor)
 
     in_data.retain_grad()
